+++ comb_spec_searcher/comb_spec_searcher.py
@@ -10,6 +10,7 @@ from typing import (
     Dict,
     Generic,
     Iterator,
+    List,
     Optional,
     Sequence,
     Set,
@@ -58,6 +59,11 @@ if platform.python_implementation() == "CPython":
 warnings.simplefilter("once", Warning)
 
 logzero.loglevel(logging.INFO)
+
+Specification = Tuple[
+    List[Tuple[CombinatorialClassType, AbstractStrategy]],
+    List[List[CombinatorialClassType]],
+]
 
 
 class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
@@ -588,26 +594,12 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
 
         max_expansion_time = 0
         expanding = True
-        last_label = None
+
         while expanding:
-            expansion_start = time.time()
-            for label, strategies, inferral in self._labels_to_expand():
-                if label != last_label:
-                    comb_class = self.classdb.get_class(label)
-                    last_label = label
-                if not self.ruledb.is_verified(label):
-                    self._expand(comb_class, label, strategies, inferral)
-                if time.time() - expansion_start > max_expansion_time:
-                    break
-                if (
-                    status_update is not None
-                    and time.time() - status_start > status_update
-                ):
-                    self._log_status(auto_search_start, status_update)
-                    status_start = time.time()
-            else:
-                expanding = False
-                logger.info("No more classes to expand.", extra=self.logger_kwargs)
+            expanding, status_start = self._expand_classes_for(
+                max_expansion_time, status_update, status_start, auto_search_start
+            )
+
             spec_search_start = time.time()
             logger.debug("Searching for specification.", extra=self.logger_kwargs)
             specification = self.get_specification(
@@ -635,6 +627,67 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                 extra=self.logger_kwargs,
             )
 
+    def _auto_search_rules(self) -> Specification:
+        """
+        A basic auto search for returning equivalence paths and rules.
+
+        This method is used by CombinatorialSpecification for expanding
+        verified classes.
+
+        Will raise SpecificationNotFound error if no specification is found
+        after running out of classes to expand.
+        """
+        status_start = time.time()
+        status_update = None  # this prevents status updates happening
+        auto_search_start = time.time()
+        max_expansion_time: float = 0
+        expanding = True
+        while expanding:
+            expanding, status_start = self._expand_classes_for(
+                max_expansion_time, status_update, status_start, auto_search_start
+            )
+            spec_search_start = time.time()
+            spec = self._get_specification_rules(
+                0.01 * (time.time() - auto_search_start)
+            )
+            if spec is not None:
+                return spec
+            max_expansion_time = min(0.01 * (time.time() - spec_search_start), 3600)
+        raise SpecificationNotFound
+
+    def _expand_classes_for(
+        self,
+        expansion_time: float,
+        status_update: Optional[int],
+        status_start: float,
+        auto_search_start: float,
+    ) -> Tuple[bool, float]:
+        """
+        Will expand classes for `expansion_time` seconds.
+
+        It will return a pair (bool, time), where the bool is True if there are
+        more classes to expand, False otherwise. The `time` is the initial time
+        for checking whether to post a status update.
+        """
+        expansion_start = time.time()
+        last_label = None
+        expanding = True
+        for label, strategies, inferral in self._labels_to_expand():
+            if label != last_label:
+                comb_class = self.classdb.get_class(label)
+                last_label = label
+            if not self.ruledb.is_verified(label):
+                self._expand(comb_class, label, strategies, inferral)
+            if time.time() - expansion_start > expansion_time:
+                break
+            if status_update is not None and time.time() - status_start > status_update:
+                self._log_status(auto_search_start, status_update)
+                status_start = time.time()
+        else:
+            expanding = False
+            logger.info("No more classes to expand.", extra=self.logger_kwargs)
+        return expanding, status_start
+
     @cssmethodtimer("get specification")
     def get_specification(
         self,
@@ -647,8 +700,36 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
 
         The minimization_time_limit only applies when smallest is false.
 
-        The function will raise a SpecificationNotFound if no such
-        CombinatorialSpecification exists in the universe.
+        The function will return None if no such CombinatorialSpecification
+        exists in the universe.
+        """
+        spec = self._get_specification_rules(minimization_time_limit, smallest)
+        if spec is None:
+            return None
+        start_class = self.classdb.get_class(self.start_label)
+        strategies, comb_class_eqv_paths = spec
+        logger.info(
+            "Creating a specification", extra=self.logger_kwargs,
+        )
+        return CombinatorialSpecification(
+            start_class,
+            strategies,
+            comb_class_eqv_paths,
+            expand_verified=expand_verified,
+        )
+
+    @cssmethodtimer("get specification")
+    def _get_specification_rules(
+        self, minimization_time_limit: float = 10, smallest: bool = False,
+    ) -> Optional[Specification]:
+        """
+        Returns the equivalence paths needed to create a
+        CombinatorialSpecification, if one exists.
+
+        The minimization_time_limit only applies when smallest is false.
+
+        The function will return None if no such CombinatorialSpecification
+        exists in the universe.
         """
         try:
             if smallest:
@@ -665,16 +746,8 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                 )
         except SpecificationNotFound:
             return None
-        start_class = self.classdb.get_class(self.start_label)
-        comb_class_eqv_paths = tuple(
-            tuple(map(self.classdb.get_class, path)) for path in eqv_paths
-        )
-        logger.info(
-            "Creating a specification", extra=self.logger_kwargs,
-        )
-        return CombinatorialSpecification(
-            start_class,
-            [(self.classdb.get_class(label), rule) for label, rule in rules],
-            comb_class_eqv_paths,
-            expand_verified=expand_verified,
-        )
+        comb_class_eqv_paths = [
+            list(map(self.classdb.get_class, path)) for path in eqv_paths
+        ]
+        strategies = [(self.classdb.get_class(label), rule) for label, rule in rules]
+        return strategies, comb_class_eqv_paths