+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `expand_verified` flag to the initialiser of
   `CombinatorialSpecificationSearcher` which will expand verified classes using
   the pack passed to the `CombinatorialSpecificationSearcher`
+- It is now possible to get all the combinatorial classes contained in a specification
+  with the `comb_classes` method.
 
 ### Changed
 - When expanding verified nodes in a specification, the search now uses