+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   `CombinatorialSpecification`
 - the `get_rule` method on `CombinatorialSpecification` can also take a label
   as a key
+- `rule_from_equivalence_rule_dict` to `RuleDBBase`
 - It is now possible to get all the combinatorial classes contained in a specification
   with the `comb_classes` method.