+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `expand_verified` flag to the initialiser of
   `CombinatorialSpecificationSearcher` which will expand verified classes using
   the pack passed to the `CombinatorialSpecificationSearcher`
+- `rule_from_equivalence_rule_dict` to `RuleDBBase`
 - It is now possible to get all the combinatorial classes contained in a specification
   with the `comb_classes` method.